requirements.txt
flask_restful==0.3.9
selenium==4.21.0
webdriver_manager==3.8.4
